scaling values up by 100

+++ project/rules.h
@@ -123,7 +123,7 @@ public:
     static bool stalemate(Colour c, const Board &board, const Move &previousMove);
 
     // returns a valuation for a move for a computer of colour c given level
-    static float evalMove(int level, Colour c, Board &board, const Move &move, const Move &previousMove);
+    static int evalMove(int level, Colour c, Board &board, const Move &move, const Move &previousMove);
 };
 
 #endif
